requirements.txt
Django
django-allauth
django-cors-headers
django-environ
django-filter
django-mssql-backend
djangorestframework
pillow
pymssql
pyodbc
mssql-django